Move setting up rollup back into soak-testing crate with binaries
