+++ crates/utils/sov-soak-testing-lib/src/lib.rs
@@ -1,5 +1,4 @@
 use std::collections::HashMap;
-use std::path::PathBuf;
 use std::sync::Arc;
 use std::time::Duration;
 
@@ -12,25 +11,12 @@ use sov_modules_api::macros::config_value;
 use sov_modules_api::prelude::arbitrary::{self, Unstructured};
 use sov_modules_api::prelude::tracing;
 use sov_modules_api::transaction::TxDetails;
-use sov_modules_api::{Amount, CryptoSpec, DispatchCall, EncodeCall, PrivateKey, Runtime, Spec};
-use sov_paymaster::{
-    PayeePolicy, PayerGenesisConfig, Paymaster, PaymasterConfig, PaymasterPolicyInitializer,
-    SafeVec,
-};
-use sov_sequencer::preferred::PreferredSequencerConfig;
-use sov_sequencer::SequencerKindConfig;
+use sov_modules_api::{CryptoSpec, DispatchCall, EncodeCall, PrivateKey, Runtime, Spec};
 use sov_synthetic_load::CallMessageDiscriminants::{
     ReadAndSetHeavyState, ReadAndSetManyIndividualValues, RunCPUHeavyOperation,
 };
 use sov_synthetic_load::SyntheticLoad;
-use sov_test_utils::runtime::genesis::zk::config::HighLevelZkGenesisConfig;
-use sov_test_utils::runtime::genesis::zk::MinimalZkGenesisConfig;
-use sov_test_utils::test_rollup::{GenesisSource, RollupBuilder, TestRollup};
-use sov_test_utils::{
-    generate_runtime, RtAgnosticBlueprint, TestProver, TestSequencer, TestSpec, TestUser,
-    TransactionType, TEST_DEFAULT_MAX_FEE, TEST_DEFAULT_MAX_PRIORITY_FEE,
-    TEST_DEFAULT_USER_BALANCE,
-};
+use sov_test_utils::{TransactionType, TEST_DEFAULT_MAX_FEE, TEST_DEFAULT_MAX_PRIORITY_FEE};
 use sov_transaction_generator::generators::bank::harness_interface::BankHarness;
 use sov_transaction_generator::generators::bank::BankMessageGenerator;
 use sov_transaction_generator::generators::basic::{
@@ -49,23 +35,7 @@ pub const DEFAULT_BLOCK_PRODUCING_CONFIG: BlockProducingConfig = BlockProducingC
     block_time_ms: DEFAULT_BLOCK_TIME_MS,
 };
 
-pub type TestRT = TestRuntime<TestSpec>;
-pub type RollupBlueprint = RtAgnosticBlueprint<TestSpec, TestRT>;
-pub type TestRollupBuilder = RollupBuilder<RollupBlueprint, PathBuf>;
-
 pub const DEFAULT_FINALIZATION_BLOCKS: u32 = 5;
-
-generate_runtime! {
-    name: TestRuntime,
-    modules: [paymaster: Paymaster<S>, synthetic_load: SyntheticLoad<S>],
-    operating_mode: sov_modules_api::runtime::OperatingMode::Zk,
-    minimal_genesis_config_type: MinimalZkGenesisConfig<S>,
-    gas_enforcer: paymaster: Paymaster<S>,
-    runtime_trait_impl_bounds: [],
-    kernel_type: sov_kernels::soft_confirmations::SoftConfirmationsKernel<'a, S>,
-    auth_type: sov_modules_api::capabilities::RollupAuthenticator<S, Self>,
-    auth_call_wrapper: |call| call,
-}
 
 pub const BUFFER_SIZE: usize = 100_000;
 // The minimum randomness needed to guarantee successful transaction generation
@@ -198,107 +168,6 @@ pub fn setup_harness<R: Runtime<S> + Clone, S: Spec>(rng_salt: u128) -> TestGene
         target_buffer_size: BUFFER_SIZE,
         salt: rng_salt,
     }
-}
-
-pub struct Setup {
-    /// A user who is pre-registered as a payer for [`Setup::sequencer`].
-    #[allow(dead_code)]
-    pub paymaster: TestUser<TestSpec>,
-    /// The pre-registered sequencer
-    pub sequencer: TestSequencer<TestSpec>,
-    /// The pre-registered prover
-    pub prover: TestProver<TestSpec>,
-    #[allow(missing_docs)]
-    pub genesis_config: GenesisConfig<TestSpec>,
-}
-
-pub fn setup_roles_and_config() -> Setup {
-    let mut genesis_config = HighLevelZkGenesisConfig::generate();
-
-    let sequencer = genesis_config.initial_sequencer.clone();
-    let prover = genesis_config.initial_prover.clone();
-    let paymaster = TestUser::generate(
-        TEST_DEFAULT_USER_BALANCE
-            .checked_mul(Amount::new(10))
-            .unwrap(),
-    );
-    genesis_config
-        .additional_accounts_mut()
-        .push(paymaster.clone());
-
-    let users: Vec<TestUser<TestSpec>> = vec![TestUser::generate_with_default_balance(); 20];
-
-    genesis_config.additional_accounts_mut().extend(users);
-    let genesis_config = GenesisConfig::from_minimal_config(
-        genesis_config.into(),
-        PaymasterConfig {
-            payers: [PayerGenesisConfig {
-                payer_address: paymaster.address(),
-                policy: PaymasterPolicyInitializer {
-                    default_payee_policy: PayeePolicy::Allow {
-                        max_fee: None,
-                        gas_limit: None,
-                        max_gas_price: None,
-                        transaction_limit: None,
-                    },
-                    payees: SafeVec::new(),
-                    authorized_sequencers: sov_paymaster::AuthorizedSequencers::All,
-                    authorized_updaters: [paymaster.address()].as_ref().try_into().unwrap(),
-                },
-                sequencers_to_register: [sequencer.da_address].as_ref().try_into().unwrap(),
-            }]
-            .as_ref()
-            .try_into()
-            .unwrap(),
-        },
-        (),
-    );
-    Setup {
-        paymaster,
-        sequencer,
-        prover,
-        genesis_config,
-    }
-}
-
-pub async fn setup_rollup(
-    storage_path: PathBuf,
-    axum_port: u16,
-    setup: Setup,
-    db_connection_url: Option<String>,
-) -> TestRollup<RollupBlueprint, PathBuf> {
-    let rollup_builder = TestRollupBuilder::new_with_storage_path(
-        GenesisSource::CustomParams(setup.genesis_config.clone().into_genesis_params()),
-        DEFAULT_BLOCK_PRODUCING_CONFIG,
-        DEFAULT_FINALIZATION_BLOCKS,
-        storage_path,
-        false,
-    )
-    .set_config(|config| {
-        config.telegraf_address = sov_metrics::MonitoringConfig::standard().telegraf_address;
-        config.automatic_batch_production = true;
-        config.rollup_prover_config = None;
-        config.sequencer_config = SequencerKindConfig::Preferred(PreferredSequencerConfig {
-            minimum_profit_per_tx: 0,
-            postgres_connection_string: db_connection_url,
-            ..Default::default()
-        });
-        config.prover_address = setup.prover.user_info.address().to_string();
-        config.aggregated_proof_block_jump = 3;
-        config.axum_port = axum_port;
-        if let SequencerKindConfig::Preferred(preferred_sequencer_config) =
-            &mut config.sequencer_config
-        {
-            preferred_sequencer_config.batch_execution_time_limit_millis = 400;
-        }
-    })
-    .set_da_config(|da_config| {
-        da_config.sender_address = setup.sequencer.da_address;
-    });
-    rollup_builder
-        .start()
-        .await
-        .expect("Impossible to start rollup")
 }
 
 /// The passed client is responsible for handling timeouts (otherwise calls can block).
