+++ examples/demo-rollup/sov-soak-testing/src/lib.rs
@@ -3,8 +3,24 @@ use sov_celestia_adapter::verifier::CelestiaSpec;
 use sov_mock_da::{BlockProducingConfig, MockDaSpec};
 use sov_mock_zkvm::MockZkvm;
 use sov_modules_api::configurable_spec::ConfigurableSpec;
+use sov_modules_api::Amount;
+use sov_paymaster::{
+    PayeePolicy, PayerGenesisConfig, Paymaster, PaymasterConfig, PaymasterPolicyInitializer,
+    SafeVec,
+};
 use sov_rollup_interface::execution_mode::Native;
+use sov_sequencer::preferred::PreferredSequencerConfig;
+use sov_sequencer::SequencerKindConfig;
 pub use sov_soak_testing_lib::*;
+use sov_synthetic_load::SyntheticLoad;
+use sov_test_utils::runtime::genesis::zk::config::HighLevelZkGenesisConfig;
+use sov_test_utils::runtime::genesis::zk::MinimalZkGenesisConfig;
+use sov_test_utils::test_rollup::{GenesisSource, RollupBuilder, TestRollup};
+use sov_test_utils::{
+    generate_runtime, RtAgnosticBlueprint, TestProver, TestSequencer, TestSpec, TestUser,
+    TEST_DEFAULT_USER_BALANCE,
+};
+use std::path::PathBuf;
 
 pub const DEFAULT_BLOCK_TIME_MS: u64 = 200;
 pub const DEFAULT_BLOCK_PRODUCING_CONFIG: BlockProducingConfig = BlockProducingConfig::Periodic {
@@ -12,6 +28,10 @@ pub const DEFAULT_BLOCK_PRODUCING_CONFIG: BlockProducingConfig = BlockProducingC
 };
 
 pub const DEFAULT_FINALIZATION_BLOCKS: u32 = 5;
+
+pub type TestRT = TestRuntime<TestSpec>;
+pub type RollupBlueprint = RtAgnosticBlueprint<TestSpec, TestRT>;
+pub type TestRollupBuilder = RollupBuilder<RollupBlueprint, PathBuf>;
 
 // Celestia
 pub type CelestiaRollupSpec =
@@ -22,3 +42,116 @@ pub type DemoCelestiaRT = demo_stf::runtime::Runtime<CelestiaRollupSpec>;
 pub type MockDemoRollupSpec =
     ConfigurableSpec<MockDaSpec, MockZkvm, MockZkvm, MultiAddressEvm, Native>;
 pub type DemoMockRT = demo_stf::runtime::Runtime<MockDemoRollupSpec>;
+
+generate_runtime! {
+    name: TestRuntime,
+    modules: [paymaster: Paymaster<S>, synthetic_load: SyntheticLoad<S>],
+    operating_mode: sov_modules_api::runtime::OperatingMode::Zk,
+    minimal_genesis_config_type: MinimalZkGenesisConfig<S>,
+    gas_enforcer: paymaster: Paymaster<S>,
+    runtime_trait_impl_bounds: [],
+    kernel_type: sov_kernels::soft_confirmations::SoftConfirmationsKernel<'a, S>,
+    auth_type: sov_modules_api::capabilities::RollupAuthenticator<S, Self>,
+    auth_call_wrapper: |call| call,
+}
+
+pub struct Setup {
+    /// A user who is pre-registered as a payer for [`Setup::sequencer`].
+    #[allow(dead_code)]
+    pub paymaster: TestUser<TestSpec>,
+    /// The pre-registered sequencer
+    pub sequencer: TestSequencer<TestSpec>,
+    /// The pre-registered prover
+    pub prover: TestProver<TestSpec>,
+    #[allow(missing_docs)]
+    pub genesis_config: GenesisConfig<TestSpec>,
+}
+
+pub fn setup_roles_and_config() -> Setup {
+    let mut genesis_config = HighLevelZkGenesisConfig::generate();
+
+    let sequencer = genesis_config.initial_sequencer.clone();
+    let prover = genesis_config.initial_prover.clone();
+    let paymaster = TestUser::generate(
+        TEST_DEFAULT_USER_BALANCE
+            .checked_mul(Amount::new(10))
+            .unwrap(),
+    );
+    genesis_config
+        .additional_accounts_mut()
+        .push(paymaster.clone());
+
+    let users: Vec<TestUser<TestSpec>> = vec![TestUser::generate_with_default_balance(); 20];
+
+    genesis_config.additional_accounts_mut().extend(users);
+    let genesis_config = GenesisConfig::from_minimal_config(
+        genesis_config.into(),
+        PaymasterConfig {
+            payers: [PayerGenesisConfig {
+                payer_address: paymaster.address(),
+                policy: PaymasterPolicyInitializer {
+                    default_payee_policy: PayeePolicy::Allow {
+                        max_fee: None,
+                        gas_limit: None,
+                        max_gas_price: None,
+                        transaction_limit: None,
+                    },
+                    payees: SafeVec::new(),
+                    authorized_sequencers: sov_paymaster::AuthorizedSequencers::All,
+                    authorized_updaters: [paymaster.address()].as_ref().try_into().unwrap(),
+                },
+                sequencers_to_register: [sequencer.da_address].as_ref().try_into().unwrap(),
+            }]
+            .as_ref()
+            .try_into()
+            .unwrap(),
+        },
+        (),
+    );
+    Setup {
+        paymaster,
+        sequencer,
+        prover,
+        genesis_config,
+    }
+}
+
+pub async fn setup_rollup(
+    storage_path: PathBuf,
+    axum_port: u16,
+    setup: Setup,
+    db_connection_url: Option<String>,
+) -> TestRollup<RollupBlueprint, PathBuf> {
+    let rollup_builder = TestRollupBuilder::new_with_storage_path(
+        GenesisSource::CustomParams(setup.genesis_config.clone().into_genesis_params()),
+        sov_soak_testing_lib::DEFAULT_BLOCK_PRODUCING_CONFIG,
+        sov_soak_testing_lib::DEFAULT_FINALIZATION_BLOCKS,
+        storage_path,
+        false,
+    )
+    .set_config(|config| {
+        config.telegraf_address = sov_metrics::MonitoringConfig::standard().telegraf_address;
+        config.automatic_batch_production = true;
+        config.rollup_prover_config = None;
+        config.sequencer_config = SequencerKindConfig::Preferred(PreferredSequencerConfig {
+            minimum_profit_per_tx: 0,
+            postgres_connection_string: db_connection_url,
+            ..Default::default()
+        });
+        config.prover_address = setup.prover.user_info.address().to_string();
+        config.aggregated_proof_block_jump = 3;
+        config.axum_port = axum_port;
+        if let SequencerKindConfig::Preferred(preferred_sequencer_config) =
+            &mut config.sequencer_config
+        {
+            preferred_sequencer_config.batch_execution_time_limit_millis = 400;
+        }
+    })
+    .set_da_config(|da_config| {
+        da_config.sender_address = setup.sequencer.da_address;
+    });
+    rollup_builder
+        .start()
+        .await
+        .expect("Impossible to start rollup")
+}
